add security level and temp peer bindings
The SSL_get_peer_tmp_key bindings are implemented as a macro so that is
why it looks very ugly.

I also needed access to set the security level, so added that as a
bindings.

I am very uncertain about what the security level of these various
things is supposed to be.